requirements.txt
fastapi
uvicorn
pydantic-settings
sqlalchemy 
psycopg2-binary
httpie
multidict